+++ docs/src/jacdac/useEventSubscription.ts
@@ -1,7 +1,7 @@
-import { EventEmitter } from "../../../src/dom/eventemitter";
+import { JDNode } from "../../../src/dom/eventemitter";
 import { useState, useEffect } from "react";
 
-function useEventSubscription<T>(eventEmitter: EventEmitter, eventName: string): T {
+function useEventSubscription<T>(eventEmitter: JDNode, eventName: string): T {
     const [value, setValue] = useState<T>(undefined)
     useEffect(() => eventEmitter.subscribe<T>(eventName, value => setValue(value))
         , [eventEmitter, eventName])
