+++ src/dom/jacdac.ts
@@ -4,7 +4,6 @@ export * from './buffer'
 export * from './msgpack'
 export * from './struct'
 export * from './observable'
-export * from './eventemitter'
 export * from './packet'
 export * from './device'
 export * from './service'
